Uudet paketit logic ja domain joihin luokat App, Reference ja Article
