+++ src/main/java/domain/Reference.java
@@ -1,0 +1,5 @@
+package domain;
+
+public class Reference {
+    
+}
